chore: 특정 domain 관련된 exception 은 domain 패키지에 배치함 (변경도 적기 때문임)
