+++ src/main/java/com/example/freshcart/order/domain/exception/OrderItemNotFoundException.java
@@ -1,19 +1,20 @@
-package com.example.freshcart.product.infrastructure.exception;
+package com.example.freshcart.order.domain.exception;
 
 import com.example.freshcart.global.exception.BaseException;
 import org.springframework.http.HttpStatus;
 
-public class NotSellerException extends BaseException {
+public class OrderItemNotFoundException extends BaseException {
 
   // static 변수 할당으로 인스턴스를 생성하지 않고도 사용 가능하게 함. 변수가 super() 생성자에서도 참조되기 위함.
-  private static final String message = "판매자만 이용 가능합니다";
+  private static final String message = "상품이 변경되었습니다. 다시 주문해주세요";
 
-  public NotSellerException() {
-    super(HttpStatus.UNAUTHORIZED, message);
+  public OrderItemNotFoundException() {
+    super(HttpStatus.NOT_FOUND, message);
   }
 
   @Override
   public String getMessage() {
     return message;
+
   }
 }
