+++ src/main/java/com/example/freshcart/user/domain/User.java
@@ -1,6 +1,6 @@
 package com.example.freshcart.user.domain;
 
-import com.example.freshcart.global.domain.Role;
+import com.example.freshcart.authentication.Role;
 
 /**
  * NotNull + Valid(@RequestBody와 함께)로 필드가 입력되었는지 점검 .
